Apps 1662 attributes

+++ tests/unittest/export_aws/test_aws_hooks.py
@@ -2,7 +2,7 @@ import os
 import unittest
 import dtlpy as dl
 
-from modules.aws_hooks import HookIntegrationAWS
+from modules.aws_hooks import AWSExport
 
 
 class Node:
@@ -12,10 +12,18 @@ class Node:
 
 class TestRunner(unittest.TestCase):
     def setUp(self):
-        self.runner = HookIntegrationAWS(integration_name='')
-        self.original_item = dl.items.get(item_id='658ae4cd160fb30cdebf1156')
-        self.original_annotations = self.original_item.annotations.list()
+        item_id = ""
+        bucket_name = ""
+        region_name = ""
         remote_filepath = "/clones/1.jpg"
+
+        # Connect .env file
+        if os.environ.get("AWS_INTEGRATION") is None:
+            raise ValueError("Missing AWS service integration.")
+        self.runner = AWSExport()
+        self.original_item = dl.items.get(item_id=item_id)
+        self.original_annotations = self.original_item.annotations.list()
+
         try:
             item = self.original_item.dataset.items.get(filepath=remote_filepath)
             item.delete()
@@ -24,7 +32,10 @@ class TestRunner(unittest.TestCase):
 
         self.item = self.original_item.clone(remote_filepath=remote_filepath)
         self.context = dl.Context()
-        self.context._node = Node(metadata={'customNodeConfig': {'bucket_name': 'micha-storage'}})
+        self.context._node = Node(metadata={'customNodeConfig': {
+            'bucket_name': bucket_name,
+            'region_name': region_name
+        }})
 
     def test_export_annotations(self):
         self.runner.export_annotation(item=self.item, context=self.context)
